Build majority of logic into Client class
